+++ NEWS.md
@@ -1,3 +1,18 @@
+# MCView 0.2.40 
+
+* Show metadata in markers heatmap tooltip.
+* Added search to cell type selectors.
+* Reset on apply by default in `Annotate` tab.
+
+# MCView 0.2.39
+
+* Added `log2` option to the scatter plots.
+* Fix: 'Stroke' was not working in the 2D projection plots.
+
+# MCView 0.2.38
+
+* Show GeneCards buttons also when Show-Correlations is off.
+
 # MCView 0.2.36
 
 * Fix: Metadata/Gene correlations were not shown.